refactor client controllers and views

+++ app/Http/Controllers/ClientController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\ClientStoreRequest;
 use App\Http\Traits\SettingTrait;
 use App\Models\client;
 use App\Models\country;
@@ -17,7 +18,7 @@ class ClientController extends Controller
     public function index()
     {
         $countries = country::all();
-        $clients = client::with('country')->get();
+        $clients = client::get();
 
         return view('backend.client.index', compact('clients', 'countries'));
     }
@@ -27,32 +28,32 @@ class ClientController extends Controller
         //
     }
 
-    public function store(Request $request, ToastrFactory $flasher)
+    public function store(ClientStoreRequest $request, ToastrFactory $flasher)
     {
-        //dd($request);
+        //  dd($request);
         try {
             $client = new client();
             $client->name = ['en' => $request->name_en, 'ar' => $request->name];
             $client->phone = $request->phone;
             $client->address = $request->address;
             $client->country_id = $request->country_id;
+            $client->code = $request->code;
+            $client->balance = $request->balance;
             $client->save();
             $flasher->AddSuccess(trans('general.add_msg'));
 
-            return redirect('client');
+            return redirect()->route('client_index');
         } catch (\Exception $e) {
             return redirect()
                 ->back()
-                ->withErrors(['error' => $e->getMessage()]);
+                ->withErrors(['error' => $e->getMessage()])->withInput();
         }
     }
 
     public function show($client)
     {
         $data = MoneyTreasary::where('client_id', $client)->with('client')->orderby('payed_at', 'asc')->get();
-        //        $data['sales'] = salesinv::where('client_id', $client->id)->get();
-        // return $data;
-        //return $data['client'];
+
         return view('backend.client.show', compact('data'));
     }
 
@@ -63,17 +64,17 @@ class ClientController extends Controller
 
     public function update(Request $request, ToastrFactory $flasher)
     {
+        //  return $request;
         try {
             $client = Client::findorfail($request->id);
-            // $client = new client();
             $client->name = ['en' => $request->name_en, 'ar' => $request->name];
             $client->phone = $request->phone;
             $client->address = $request->address;
-            $client->country_id = $request->country_id;
+            $client->country_id = ($request->country_id) ? $request->country_id : $client->country_id;
             $client->save();
             $flasher->AddInfo(trans('general.edit_msg'));
 
-            return redirect('client');
+            return redirect()->route('client_index');
         } catch (\Exception $e) {
             return redirect()
                 ->back()
@@ -84,10 +85,15 @@ class ClientController extends Controller
     public function destroy(Request $request, ToastrFactory $flasher)
     {
         try {
-            client::destroy($request->id);
-            $flasher->AddSuccess(trans('general.delete_msg'));
+            $sales_invs = salesinv::where('client_id', $request->id)->count();
+            if ($sales_invs == 0) {
+                client::destroy($request->id);
+                $flasher->AddSuccess(trans('general.delete_msg'));
+            } else {
+                $flasher->AddError(trans('client.canotdelete'));
+            }
 
-            return redirect('client');
+            return redirect()->route('client_index');
         } catch (\Exception $e) {
             return redirect()
                 ->back()
@@ -98,10 +104,9 @@ class ClientController extends Controller
     public function print($id)
     {
         try {
-            $client = client::findorfail($id);
-            $sales = salesinv::where('client_id', $id)->get();
-
-            return view('backend.client.print_client_balance', compact('sales', 'client'), $this->GetData());
+            $data = MoneyTreasary::where('client_id', $id)->with('client')->orderby('payed_at', 'asc')->get();
+            // return $data[0]->client->totalBalance();
+            return view('backend.client.print_client_balance', compact('data'), $this->GetData());
         } catch (\Exception $e) {
             return redirect()->back()
                 ->withErrors(['error' => $e->getMessage()]);
